abstract base invite banner
